GsFileBrowserFragment+ListAdapter: move FileObserver into adapter, use makeDebounced
Addresses review on PR #2823 from @harshad1 and @gsantner:

  - "We have a system to debounce with a delay (search for makeDebounced)"
  - "IMO we should try to move this into the adapter. That would simplify several things."

FileObserver state now lives in GsFileBrowserListAdapter rather than the
fragment. The adapter already owns _currentFolder, navigation, and the
reload pipeline, so co-locating the watcher removes the need for the
fragment to mirror that state or coordinate lifecycle.

Wiring:

  - rebindFolderObserver() is called from loadFolder() only when
    folderChanged is true. This means folder navigation and back-stack
    pops rebind the watch automatically, while pull-to-refresh and
    same-folder reloads (including auto-refresh-triggered reloads) do
    not churn the kernel inotify watch.

  - Guarded on folder.isDirectory() && folder.canRead() so the virtual
    folders (Recent/Popular/Favourite/AppData/storage-emulated) and
    permission-denied scoped-storage paths skip cleanly. FileObserver
    events on these would be meaningless since their listings are
    synthesised, not read from disk.

  - onEvent dispatches through TextViewUtils.makeDebounced(300, ...)
    instead of the hand-rolled Handler.postDelayed pair. Coalesces
    bursts (50-file sync, multi-select delete) to a single reload.
    makeDebounced posts on the main-thread Handler, so the actual
    reloadCurrentFolder() runs on the UI thread regardless of which
    FileObserver internal thread fired the event.

  - onDetachedFromRecyclerView() stops the watch. The adapter is
    constructed fresh per fragment view, so the natural pairing is
    attach-to-RecyclerView (first loadFolder binds via folderChanged)
    and detach-from-RecyclerView (releases the kernel watch).

Net effect on the fragment: it loses the FileObserver field, the
Handler/Looper plumbing, onPause(), and the start/stop helpers. The
GsFileBrowserDialog form, which shares the adapter, inherits the
auto-refresh for free.
